requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==21.4.0
black==22.1.0
blobfile==1.2.8
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cog==0.0.3
cycler==0.11.0
einops==0.3.2
filelock==3.4.2
fonttools==4.29.0
frozenlist==1.3.0
fsspec==2022.1.0
ftfy==6.0.3
future==0.18.2
-e git+https://git@github.com/openai/glide-text2im.git@9cc8e563851bd38f5ddb3e305127192cb0f02f5c#egg=glide_text2im
google-auth==2.5.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
-e git+https://git@github.com/openai/guided-diffusion.git@27c20a8fab9cb472df5d6bdd6c8d11c8f430b924#egg=guided_diffusion
idna==3.3
importlib-metadata==4.10.1
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
mpi4py==3.1.3
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.21.5
oauthlib==3.2.0
omegaconf==2.1.1
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.14.0
pyDeprecate==0.3.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-lightning==1.5.7
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
smmap==5.0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
toml==0.10.2
tomli==2.0.0
torch==1.10.1
torchmetrics==0.7.0
tornado==6.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
xmltodict==0.12.0
yarl==1.7.2
zipp==3.7.0
